requirements.txt
pdf2image==1.17.0
pdfplumber==0.11.4
pytesseract==0.3.13
python-dotenv==1.0.1
streamlit==1.32.0
google.generativeai
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.64.0
